windows/requirements.txt
keyboard>=0.13.5
numpy>=1.24.0
Pillow>=10.0.0
pyperclip>=1.9.0
pystray>=0.19.5
qwen-asr>=0.0.6
sounddevice>=0.4.7
torch @ https://download.pytorch.org/whl/cu128/torch-2.9.1%2Bcu128-cp311-cp311-win_amd64.whl
transformers>=4.50.0

# Testing only
pytest>=8.0.0
